(layout) feat: add default layout for app
