+++ src/components/sidebar/SidebarTools.vue
